Move render tests to separate module
They're distinct tests, and the load-speed is not
large enough to matter.
